feat(DTFS2-7225): remove isTfmFacilityEndDateFeatureFlagEnabled
