+++ libs/common/src/helpers/is-feature-flag-enabled.test.ts
@@ -2,7 +2,6 @@ import {
   isFeeRecordCorrectionFeatureFlagEnabled,
   isPortalFacilityAmendmentsFeatureFlagEnabled,
   isTfmDealCancellationFeatureFlagEnabled,
-  isTfmFacilityEndDateFeatureFlagEnabled,
   isTfmPaymentReconciliationFeatureFlagEnabled,
   isTfmSsoFeatureFlagEnabled,
 } from './is-feature-flag-enabled';
@@ -11,7 +10,6 @@ import { withBooleanFeatureFlagTests } from './with-boolean-feature-flag.tests';
 describe('is-feature-flag-enabled helpers', () => {
   withBooleanFeatureFlagTests({ featureFlagName: 'FF_TFM_PAYMENT_RECONCILIATION_ENABLED', getFeatureFlagValue: isTfmPaymentReconciliationFeatureFlagEnabled });
   withBooleanFeatureFlagTests({ featureFlagName: 'FF_FEE_RECORD_CORRECTION_ENABLED', getFeatureFlagValue: isFeeRecordCorrectionFeatureFlagEnabled });
-  withBooleanFeatureFlagTests({ featureFlagName: 'FF_TFM_FACILITY_END_DATE_ENABLED', getFeatureFlagValue: isTfmFacilityEndDateFeatureFlagEnabled });
   withBooleanFeatureFlagTests({ featureFlagName: 'FF_TFM_DEAL_CANCELLATION_ENABLED', getFeatureFlagValue: isTfmDealCancellationFeatureFlagEnabled });
   withBooleanFeatureFlagTests({ featureFlagName: 'FF_PORTAL_FACILITY_AMENDMENTS_ENABLED', getFeatureFlagValue: isPortalFacilityAmendmentsFeatureFlagEnabled });
   withBooleanFeatureFlagTests({ featureFlagName: 'FF_TFM_SSO_ENABLED', getFeatureFlagValue: isTfmSsoFeatureFlagEnabled });
