+++ trade-finance-manager-api/src/v1/controllers/amendment.controller.js
@@ -1,6 +1,6 @@
 const { ObjectId } = require('mongodb');
 const { generateTfmAuditDetails } = require('@ukef/dtfs2-common/change-stream');
-const { isTfmFacilityEndDateFeatureFlagEnabled, AMENDMENT_QUERIES } = require('@ukef/dtfs2-common');
+const { AMENDMENT_QUERIES } = require('@ukef/dtfs2-common');
 const isGefFacility = require('../rest-mappings/helpers/isGefFacility');
 const api = require('../api');
 const acbs = require('./acbs.controller');
@@ -97,7 +97,7 @@ const createAmendmentTFMObject = async (amendmentId, facilityId, auditDetails) =
     tfmToAdd = await addLatestAmendmentCoverEndDate(tfmToAdd, latestCoverEndDateResponse, facilityId);
 
     let latestFacilityEndDateResponse;
-    const isFacilityEndDateEnabled = isTfmFacilityEndDateFeatureFlagEnabled() && isGefFacility(facility.facilitySnapshot.type);
+    const isFacilityEndDateEnabled = isGefFacility(facility.facilitySnapshot.type);
     if (isFacilityEndDateEnabled) {
       latestFacilityEndDateResponse = await api.getLatestCompletedAmendmentFacilityEndDate(facilityId);
       tfmToAdd = await addLatestAmendmentFacilityEndDate(tfmToAdd, latestFacilityEndDateResponse);
