+++ trade-finance-manager-ui/server/controllers/case/amendments/amendCoverEndDate.controller.js
@@ -1,5 +1,5 @@
 const { format, fromUnixTime, getUnixTime } = require('date-fns');
-const { AMENDMENT_STATUS, isTfmFacilityEndDateFeatureFlagEnabled } = require('@ukef/dtfs2-common');
+const { AMENDMENT_STATUS } = require('@ukef/dtfs2-common');
 const { HttpStatusCode } = require('axios');
 const api = require('../../../api');
 const { coverEndDateValidation } = require('./validation/amendCoverEndDateDate.validate');
@@ -112,7 +112,7 @@ const postAmendCoverEndDate = async (req, res) => {
       throw new Error('Error updating amendment with cover end date');
     }
 
-    const showFacilityEndDatePage = isTfmFacilityEndDateFeatureFlagEnabled() && facility.facilitySnapshot.isGef;
+    const showFacilityEndDatePage = facility.facilitySnapshot.isGef;
 
     return res.redirect(getNextPageUrl({ changeFacilityValue, showFacilityEndDatePage, baseUrl }));
   } catch (error) {
