+++ trade-finance-manager-ui/server/controllers/case/amendments/amendmentBankReviewDate.controller.js
@@ -1,5 +1,5 @@
 const { HttpStatusCode } = require('axios');
-const { AMENDMENT_STATUS, isTfmFacilityEndDateFeatureFlagEnabled } = require('@ukef/dtfs2-common');
+const { AMENDMENT_STATUS } = require('@ukef/dtfs2-common');
 const { format, parseISO } = require('date-fns');
 const api = require('../../../api');
 const { bankReviewDateValidation } = require('./validation/amendmentBankReviewDate.validate');
@@ -22,7 +22,7 @@ const getAmendmentBankReviewDate = async (req, res) => {
     return res.redirect('/not-found');
   }
 
-  const isFacilityEndDateEnabled = isTfmFacilityEndDateFeatureFlagEnabled() && facility.facilitySnapshot.isGef;
+  const isFacilityEndDateEnabled = facility.facilitySnapshot.isGef;
 
   if (!changeCoverEndDate || !isFacilityEndDateEnabled) {
     return res.redirect(`/case/${amendment.dealId}/facility/${facilityId}/amendment/${amendmentId}/amendment-options`);
@@ -69,7 +69,7 @@ const postAmendmentBankReviewDate = async (req, res) => {
   const { error, bankReviewDate } = bankReviewDateValidation({ day, month, year }, coverStartDate);
 
   if (error?.fields) {
-    const isFacilityEndDateEnabled = isTfmFacilityEndDateFeatureFlagEnabled() && facility.facilitySnapshot.isGef;
+    const isFacilityEndDateEnabled = facility.facilitySnapshot.isGef;
 
     const isEditable = amendment.status === AMENDMENT_STATUS.IN_PROGRESS && amendment.changeCoverEndDate && isFacilityEndDateEnabled;
