+++ trade-finance-manager-ui/server/controllers/case/amendments/amendmentIsUsingFacilityEndDate.controller.js
@@ -1,4 +1,4 @@
-const { AMENDMENT_STATUS, isTfmFacilityEndDateFeatureFlagEnabled } = require('@ukef/dtfs2-common');
+const { AMENDMENT_STATUS } = require('@ukef/dtfs2-common');
 const { HttpStatusCode } = require('axios');
 const api = require('../../../api');
 const { isUsingFacilityEndDateValidation } = require('./validation/amendmentIsUsingFacilityEndDate.validate');
@@ -22,7 +22,7 @@ const getAmendmentIsUsingFacilityEndDate = async (req, res) => {
     return res.redirect('/not-found');
   }
 
-  const isFacilityEndDateEnabled = isTfmFacilityEndDateFeatureFlagEnabled() && facility.facilitySnapshot.isGef;
+  const isFacilityEndDateEnabled = facility.facilitySnapshot.isGef;
 
   if (!amendment.changeCoverEndDate || !isFacilityEndDateEnabled) {
     return res.redirect(`/case/${amendment.dealId}/facility/${facilityId}/amendment/${amendmentId}/amendment-options`);
@@ -62,7 +62,7 @@ const postAmendmentIsUsingFacilityEndDate = async (req, res) => {
   const { errors } = isUsingFacilityEndDateValidation(isUsingFacilityEndDate);
 
   if (errors.errorSummary?.length) {
-    const isFacilityEndDateEnabled = isTfmFacilityEndDateFeatureFlagEnabled() && facility.facilitySnapshot.isGef;
+    const isFacilityEndDateEnabled = facility.facilitySnapshot.isGef;
 
     const isEditable = amendment.status === AMENDMENT_STATUS.IN_PROGRESS && amendment.changeCoverEndDate && isFacilityEndDateEnabled;
     return res.render('case/amendments/amendment-is-using-facility-end-date.njk', {
